Changed UpdateTimeStep
