+++ plugins/native/module/vtkF3DQuakeMDLImporter.h
@@ -5,29 +5,22 @@
 
 #ifndef vtkF3DQuakeMDLImporter_h
 #define vtkF3DQuakeMDLImporter_h
-#include <vtkImporter.h>
+#include <vtkF3DImporter.h>
 #include "vtkF3DQuakeMDLImporterConstants.h"
 
-class vtkF3DQuakeMDLImporter : public vtkImporter
+class vtkF3DQuakeMDLImporter : public vtkF3DImporter
 {
 public:
   static vtkF3DQuakeMDLImporter* New();
-  //vtkTypeMacro(vtkF3DQuakeMDLImporter, vtkImporter);
 
-  void PrintSelf(ostream& os, vtkIndent indent) override;
-
-  ///@{
   /**
-   * Set/Get the file name.
+   * Set the file name.
    */
-  vtkSetMacro(FileName, std::string);
-  vtkGetMacro(FileName, std::string);
-  ///@}
+  void SetFileName(std::string fileName);
 
   /**
    * Update actors at the given time value.
    */
-
   void UpdateTimeStep(double timeValue) override;
 
   /**
@@ -43,7 +36,6 @@ public:
   ///@{
   /**
    * Enable/Disable/Get the status of specific animations
-   * Only one single animation can be enabled
    */
   void EnableAnimation(vtkIdType animationIndex) override;
   void DisableAnimation(vtkIdType animationIndex) override;
@@ -59,7 +51,6 @@ public:
 
   /**
    * Get temporal information for the currently enabled animation.
-   * Only defines timerange and ignore provided frameRate.
    */
   bool GetTemporalInformation(vtkIdType animationIndex, double frameRate, int& nbTimeSteps,
     double timeRange[2], vtkDoubleArray* timeSteps) override;
@@ -76,7 +67,6 @@ public:
 
   /**
    * Enable a specific camera.
-   * If a negative index is provided, no camera from the importer is used.
    */
   void SetCamera(vtkIdType camIndex) override;
 
@@ -88,27 +78,6 @@ protected:
   void ImportActors(vtkRenderer*) override;
   void ImportCameras(vtkRenderer*) override;
   void ImportLights(vtkRenderer*) override;
-
-  // Header definition,
-  struct mdl_header_t
-  {
-    int IDPO;
-    int version;
-    float scale[3];
-    float translation[3];
-    float boundingRadius;
-    float eyePosition[3];
-    int numSkins;
-    int skinWidth;
-    int skinHeight;
-    int numVertices;
-    int numTriangles;
-    int numFrames;
-    int syncType;
-    int stateFlags;
-    float size;
-  };
-
 
 private:
   vtkF3DQuakeMDLImporter(const vtkF3DQuakeMDLImporter&) = delete;
